refactor: 랜덤 넘버 생성 로직 분리
